Chore/ipfs upload on attachment (#67)
* IPFS upload on attachment creation

We need to upload on attachment creation rather than on submission so that the db doesn't need to store binary data as this will be unavailabe to the block indexer. Also note size has been made optional to reflect that the indexer will not have this information either but instead it is updated on fetching the resource. I will also try to get this information oportunistically in the indexer but if we don't have it null is better than the indexer failing.

* 0.6.9

* Update db docs

* 0.6.10

* Ignore unused vars with rest spread