Fix a bug whereby too many node connections are created under high load leading to 429s
Adds a test for creating 500 demandAs in parrallel and then fixs exposed issues. Broadly speaking this is caused by each controller being instantiated for each request. This was then making a ChainNode per request and thus a polkadot instance. I've made `ChainNode` a singleton now to avoid this.

In order to make tests perform sensibly I've also improved the way we were polling for state changes